Use int instead of int64 for file sizes
We are not going to use the extra bit, and String.length returns the
size as an int anyway.

+++ db/builder_db.ml
@@ -12,7 +12,7 @@ type file = Rep.file = {
   filepath : Fpath.t;
   localpath : Fpath.t;
   sha256 : Cstruct.t;
-  size : int64;
+  size : int;
 }
 
 let last_insert_rowid =
